Extract Notification to their own file
