+++ packages/core/src/result-summary.ts
@@ -15,8 +15,10 @@
  * limitations under the License.
  */
 
-import Integer, { int } from './integer'
+import Integer from './integer'
 import { NumberOrInteger } from './graph-types'
+import { util } from './internal'
+import Notification, { GqlStatusObject } from './notification'
 
 /**
  * A ResultSummary instance contains structured metadata for a {@link Result}.
@@ -366,9 +368,9 @@ class QueryStatistics {
       // To camelCase
       const camelCaseIndex = index.replace(/(-\w)/g, m => m[1].toUpperCase())
       if (camelCaseIndex in this._stats) {
-        this._stats[camelCaseIndex] = intValue(statistics[index])
+        this._stats[camelCaseIndex] = util.toNumber(statistics[index])
       } else if (camelCaseIndex === 'systemUpdates') {
-        this._systemUpdates = intValue(statistics[index])
+        this._systemUpdates = util.toNumber(statistics[index])
       } else if (camelCaseIndex === 'containsSystemUpdates') {
         this._containsSystemUpdates = statistics[index]
       } else if (camelCaseIndex === 'containsUpdates') {
@@ -419,252 +421,6 @@ class QueryStatistics {
   }
 }
 
-interface NotificationPosition {
-  offset?: number
-  line?: number
-  column?: number
-}
-
-type NotificationSeverityLevel = 'WARNING' | 'INFORMATION' | 'UNKNOWN'
-/**
- * @typedef {'WARNING' | 'INFORMATION' | 'UNKNOWN'} NotificationSeverityLevel
- */
-/**
- * Constants that represents the Severity level in the {@link Notification}
- */
-const notificationSeverityLevel: { [key in NotificationSeverityLevel]: key } = {
-  WARNING: 'WARNING',
-  INFORMATION: 'INFORMATION',
-  UNKNOWN: 'UNKNOWN'
-}
-
-Object.freeze(notificationSeverityLevel)
-const severityLevels = Object.values(notificationSeverityLevel)
-
-type NotificationCategory = 'HINT' | 'UNRECOGNIZED' | 'UNSUPPORTED' | 'PERFORMANCE' |
-'TOPOLOGY' | 'SECURITY' | 'DEPRECATION' | 'GENERIC' | 'UNKNOWN'
-/**
- * @typedef {'HINT' | 'UNRECOGNIZED' | 'UNSUPPORTED' |'PERFORMANCE' | 'TOPOLOGY' | 'SECURITY' | 'DEPRECATION' | 'GENERIC' | 'UNKNOWN' } NotificationCategory
- */
-/**
- * Constants that represents the Category in the {@link Notification}
- */
-const notificationCategory: { [key in NotificationCategory]: key } = {
-  HINT: 'HINT',
-  UNRECOGNIZED: 'UNRECOGNIZED',
-  UNSUPPORTED: 'UNSUPPORTED',
-  PERFORMANCE: 'PERFORMANCE',
-  DEPRECATION: 'DEPRECATION',
-  TOPOLOGY: 'TOPOLOGY',
-  SECURITY: 'SECURITY',
-  GENERIC: 'GENERIC',
-  UNKNOWN: 'UNKNOWN'
-}
-
-Object.freeze(notificationCategory)
-const categories = Object.values(notificationCategory)
-
-type NotificationClassification = NotificationCategory
-/**
- * @typedef {NotificationCategory} NotificationClassification
- * @experimental
- */
-/**
- * Constants that represents the Classification in the {@link GqlStatusObject}
- * @type {notificationCategory}
- * @experimental
- */
-const notificationClassification = notificationCategory
-
-/**
- * Class for Cypher notifications
- * @access public
- */
-class Notification {
-  code: string
-  title: string
-  description: string
-  severity: string
-  position: NotificationPosition | {}
-  severityLevel: NotificationSeverityLevel
-  category: NotificationCategory
-  rawSeverityLevel: string
-  rawCategory?: string
-
-  /**
-   * Create a Notification instance
-   * @constructor
-   * @param {Object} notification - Object with notification data
-   */
-  constructor (notification: any) {
-    /**
-     * The code
-     * @type {string}
-     * @public
-     */
-    this.code = notification.code
-    /**
-     * The title
-     * @type {string}
-     * @public
-     */
-    this.title = notification.title
-    /**
-     * The description
-     * @type {string}
-     * @public
-     */
-    this.description = notification.description
-    /**
-     * The raw severity
-     *
-     * Use {@link Notification#rawSeverityLevel} for the raw value or {@link Notification#severityLevel} for an enumerated value.
-     *
-     * @type {string}
-     * @public
-     * @deprecated This property will be removed in 6.0.
-     */
-    this.severity = notification.severity
-    /**
-     * The position which the notification had occur.
-     *
-     * @type {NotificationPosition}
-     * @public
-     */
-    this.position = Notification._constructPosition(notification.position)
-
-    /**
-     * The severity level
-     *
-     * @type {NotificationSeverityLevel}
-     * @public
-     * @example
-     * const { summary } = await session.run("RETURN 1")
-     *
-     * for (const notification of summary.notifications) {
-     *     switch(notification.severityLevel) {
-     *         case neo4j.notificationSeverityLevel.INFORMATION: // or simply 'INFORMATION'
-     *             console.info(`${notification.title} - ${notification.description}`)
-     *             break
-     *         case neo4j.notificationSeverityLevel.WARNING: // or simply 'WARNING'
-     *             console.warn(`${notification.title} - ${notification.description}`)
-     *             break
-     *         case neo4j.notificationSeverityLevel.UNKNOWN: // or simply 'UNKNOWN'
-     *         default:
-     *             // the raw info came from the server could be found at notification.rawSeverityLevel
-     *             console.log(`${notification.title} - ${notification.description}`)
-     *             break
-     *     }
-     * }
-     */
-    this.severityLevel = severityLevels.includes(notification.severity)
-      ? notification.severity
-      : notificationSeverityLevel.UNKNOWN
-
-    /**
-     * The severity level returned by the server without any validation.
-     *
-     * @type {string}
-     * @public
-     */
-    this.rawSeverityLevel = notification.severity
-
-    /**
-     * The category
-     *
-     * @type {NotificationCategory}
-     * @public
-     * @example
-     * const { summary } = await session.run("RETURN 1")
-     *
-     * for (const notification of summary.notifications) {
-     *     switch(notification.category) {
-     *         case neo4j.notificationCategory.QUERY: // or simply 'QUERY'
-     *             console.info(`${notification.title} - ${notification.description}`)
-     *             break
-     *         case neo4j.notificationCategory.PERFORMANCE: // or simply 'PERFORMANCE'
-     *             console.warn(`${notification.title} - ${notification.description}`)
-     *             break
-     *         case neo4j.notificationCategory.UNKNOWN: // or simply 'UNKNOWN'
-     *         default:
-     *             // the raw info came from the server could be found at notification.rawCategory
-     *             console.log(`${notification.title} - ${notification.description}`)
-     *             break
-     *     }
-     * }
-     */
-    this.category = categories.includes(notification.category)
-      ? notification.category
-      : notificationCategory.UNKNOWN
-
-    /**
-     * The category returned by the server without any validation.
-     *
-     * @type {string|undefined}
-     * @public
-     */
-    this.rawCategory = notification.category
-  }
-
-  static _constructPosition (pos: NotificationPosition): NotificationPosition {
-    if (pos == null) {
-      return {}
-    }
-    /* eslint-disable @typescript-eslint/no-non-null-assertion */
-    return {
-      offset: intValue(pos.offset!),
-      line: intValue(pos.line!),
-      column: intValue(pos.column!)
-    }
-    /* eslint-enable @typescript-eslint/no-non-null-assertion */
-  }
-}
-
-/**
- * Representation for GqlStatusObject found when executing a query.
- * <p>
- * This object represents a status of query execution.
- * This status is a superset of {@link Notification}.
- *
- * @experimental
- */
-class GqlStatusObject {
-  /**
-   * The GQLSTATUS
-   */
-  getGqlStatus (): String {
-    return ''
-  }
-
-  /**
-   * Retrieve the severity from the diagnostic record.
-   */
-  getSeverity (): NotificationSeverityLevel {
-    return notificationSeverityLevel.UNKNOWN
-  }
-
-  /**
-   * Retrieve the severity from the diagnostic record as string.
-   */
-  getRawSeverity (): String {
-    return ''
-  }
-
-  /**
-     * Retrieve the classification from the diagnostic record.
-     */
-  getClassification (): NotificationClassification {
-    return notificationClassification.UNKNOWN
-  }
-
-  /**
-   * Retrieve the classification from the diagnostic record as string
-   */
-  getRawClassification (): String {
-    return ''
-  }
-}
-
 /**
  * Class for exposing server info from a result.
  * @access public
@@ -707,16 +463,6 @@ class ServerInfo {
   }
 }
 
-function intValue (value: NumberOrInteger): number {
-  if (value instanceof Integer) {
-    return value.toNumber()
-  } else if (typeof value === 'bigint') {
-    return int(value).toNumber()
-  } else {
-    return value
-  }
-}
-
 function valueOrDefault (
   key: string,
   values: { [key: string]: NumberOrInteger } | false,
@@ -724,7 +470,7 @@ function valueOrDefault (
 ): number {
   if (values !== false && key in values) {
     const value = values[key]
-    return intValue(value)
+    return util.toNumber(value)
   } else {
     return defaultValue
   }
@@ -744,20 +490,10 @@ const queryType = {
 export {
   queryType,
   ServerInfo,
-  Notification,
   Plan,
   ProfiledPlan,
   QueryStatistics,
-  Stats,
-  notificationSeverityLevel,
-  notificationCategory,
-  notificationClassification
-}
-export type {
-  NotificationPosition,
-  NotificationSeverityLevel,
-  NotificationCategory,
-  NotificationClassification
+  Stats
 }
 
 export default ResultSummary
