+++ packages/core/test/result-summary.test.ts
@@ -18,15 +18,9 @@
 import { int } from '../src'
 import {
   ServerInfo,
-  Notification,
-  NotificationSeverityLevel,
-  NotificationCategory,
-  notificationSeverityLevel,
-  notificationCategory,
   ProfiledPlan,
   QueryStatistics,
-  Stats,
-  notificationClassification
+  Stats
 } from '../src/result-summary'
 
 import fc from 'fast-check'
@@ -59,132 +53,6 @@ describe('ServerInfo', () => {
       expect(new ServerInfo(meta, protocolVersion)).toEqual(expectedServerInfo)
     }
   )
-})
-
-describe('Notification', () => {
-  describe('.severityLevel', () => {
-    it.each(getValidSeverityLevels())('should fill severityLevel with the rawSeverityLevel equals to %s', rawSeverityLevel => {
-      const rawNotification = {
-        severity: rawSeverityLevel
-      }
-
-      const notification = new Notification(rawNotification)
-
-      expect(notification.severityLevel).toBe(rawSeverityLevel)
-      expect(notification.rawSeverityLevel).toBe(rawSeverityLevel)
-    })
-
-    it.each([
-      'UNKNOWN',
-      null,
-      undefined,
-      'I_AM_NOT_OKAY',
-      'information'
-    ])('should fill severityLevel UNKNOWN if the rawSeverityLevel equals to %s', rawSeverityLevel => {
-      const rawNotification = {
-        severity: rawSeverityLevel
-      }
-
-      const notification = new Notification(rawNotification)
-
-      expect(notification.severityLevel).toBe('UNKNOWN')
-      expect(notification.rawSeverityLevel).toBe(rawSeverityLevel)
-    })
-  })
-
-  describe('.category', () => {
-    it.each(getValidCategories())('should fill category with the rawCategory equals to %s', rawCategory => {
-      const rawNotification = {
-        category: rawCategory
-      }
-
-      const notification = new Notification(rawNotification)
-
-      expect(notification.category).toBe(rawCategory)
-      expect(notification.rawCategory).toBe(rawCategory)
-    })
-
-    it.each([
-      'UNKNOWN',
-      undefined,
-      null,
-      'DUNNO',
-      'deprecation'
-    ])('should fill category with UNKNOWN the rawCategory equals to %s', rawCategory => {
-      const rawNotification = {
-        category: rawCategory
-      }
-
-      const notification = new Notification(rawNotification)
-
-      expect(notification.category).toBe('UNKNOWN')
-      expect(notification.rawCategory).toBe(rawCategory)
-    })
-  })
-})
-
-describe('notificationSeverityLevel', () => {
-  it('should have keys equals to values', () => {
-    for (const [key, value] of Object.entries(notificationSeverityLevel)) {
-      expect(key).toEqual(value)
-    }
-  })
-
-  it('should have values assignable to NotificationSeverityLevel', () => {
-    for (const [, value] of Object.entries(notificationSeverityLevel)) {
-      const assignableValue: NotificationSeverityLevel = value
-      expect(assignableValue).toBeDefined()
-    }
-  })
-
-  it.each(getValidSeverityLevels())('should have %s as key', (severity) => {
-    const keys = Object.keys(notificationSeverityLevel)
-    expect(keys.includes(severity)).toBe(true)
-  })
-})
-
-describe('notificationCategory', () => {
-  it('should have keys equals to values', () => {
-    for (const [key, value] of Object.entries(notificationCategory)) {
-      expect(key).toEqual(value)
-    }
-  })
-
-  it('should values be assignable to NotificationCategory', () => {
-    for (const [, value] of Object.entries(notificationCategory)) {
-      const assignableValue: NotificationCategory = value
-      expect(assignableValue).toBeDefined()
-    }
-  })
-
-  it.each(getValidCategories())('should have %s as key', (category) => {
-    const keys = Object.keys(notificationCategory)
-    expect(keys.includes(category)).toBe(true)
-  })
-})
-
-describe('notificationClassification', () => {
-  it('should have keys equals to values', () => {
-    for (const [key, value] of Object.entries(notificationClassification)) {
-      expect(key).toEqual(value)
-    }
-  })
-
-  it('should values be assignable to NotificationClassification', () => {
-    for (const [, value] of Object.entries(notificationClassification)) {
-      const assignableValue: NotificationCategory = value
-      expect(assignableValue).toBeDefined()
-    }
-  })
-
-  it.each(getValidCategories())('should have %s as key', (category) => {
-    const keys = Object.keys(notificationClassification)
-    expect(keys.includes(category)).toBe(true)
-  })
-
-  it('should be notificationCategory', () => {
-    expect(notificationClassification).toBe(notificationCategory)
-  })
 })
 
 describe('ProfilePlan', () => {
@@ -383,25 +251,3 @@ describe('QueryStatistics', () => {
     })
   })
 })
-
-function getValidSeverityLevels (): NotificationSeverityLevel[] {
-  return [
-    'WARNING',
-    'INFORMATION',
-    'UNKNOWN'
-  ]
-}
-
-function getValidCategories (): NotificationCategory[] {
-  return [
-    'HINT',
-    'UNRECOGNIZED',
-    'UNSUPPORTED',
-    'PERFORMANCE',
-    'TOPOLOGY',
-    'SECURITY',
-    'DEPRECATION',
-    'GENERIC',
-    'UNKNOWN'
-  ]
-}
